% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro.R
\name{N.test}
\alias{N.test}
\title{Interval estimation and hypothesis testing function of mean when the variance of a normal population is known}
\usage{
N.test(x, mu, sigma, a = 0.05, alternative = "two.side")
}
\arguments{
\item{x}{the sample data}

\item{mu}{mean of test}

\item{sigma}{known variance}

\item{a}{the confidence level}

\item{alternative}{Unilateral or bilateral inspection}
}
\value{
interval estimation and hypothesis testing
}
\description{
Interval estimation and hypothesis testing function of mean when the variance of a normal population is known
}
\examples{
\dontrun{
  X <- 1:50 
  N.test(x=X,mu=24,sigma=9)
  }
}
