% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro.R
\name{norm.fun}
\alias{norm.fun}
\title{Find the maximum likelihood function for the normal distribution}
\usage{
norm.fun(theta, x)
}
\arguments{
\item{theta}{A parameter vector containing the mean and variance}

\item{x}{the sample data}
}
\value{
maximum likelihood estimation of the mean and variance
}
\description{
Find the maximum likelihood function for the normal distribution
}
\examples{
\dontrun{
  theta0 <- c(0,1) 
  X <- 1:50 
  result <- optim(theta0,norm.fun,x=X) 
  result
  }
}
