% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro.R
\name{Solveequation}
\alias{Solveequation}
\title{A solution of solving equation by dichotomy in fixed interval using R}
\usage{
Solveequation(f, a, b, eps = 1e-05)
}
\arguments{
\item{f}{the equation to be solved}

\item{a}{the lower bound of interval}

\item{b}{the upper bound of interval}

\item{eps}{the required precision}
}
\value{
the root of the equation
}
\description{
A solution of solving equation by dichotomy in fixed interval using R
}
\examples{
\dontrun{
  f <- function(x){
   x^3 - x - 1
   }
  Solveequation(f,1,2,1e-6)
  }
}
