\name{StatComp20001-package}
\alias{StatComp20001-package}
\alias{StatComp20001}
\docType{package}
\title{
  This package provides several functions for benchmarking R and Rcpp.
}
\description{
  A function that solves the equation by dichotomy using R, find the maximum likelihood function for the normal distribution using R, interval estimation and hypothesis testing function of mean when the variance of a normal population is known using R, and a function generating standard Laplace distribution using Rcpp are provided in this package
}
\author{
Tianxiang Ren

Maintainer: Tianxiang Ren <2111427099@qq.com>
}
\keyword{ package }

